requirements.txt
boto3==1.36.3
maturin==1.8.1
numpy==2.2.2
rasterio==1.4.3
shapely==2.0.6
toml==0.10.2
scipy==1.15.1 
pillow==11.1.0 